refactor(mcp): separate the tool contract tests from the dispatcher
